JavaFX: Mouse click and Tooltip support in Ruler
Mouse clicks in the Ruler component now operate as in the Swing UI
(except obviously that we have no Tracks yet).
Added tooltips to Ruler.  Behavior looks close enough to the Swing UI,
though it requires some interal hacks:
- The Java 8 API has no way to change the tooltip duration settings as
allowed in Swing, and the defaults are unworkable for us.  I've added
a hack (in JavaFXUIUtilities) to change these values via reflection.
Note that this *will not* work in Java 9, but 9 adds the necessary API
calls.
- The JavaFX tooltip text-update behavior is slightly different, but I
added some simple code that makes it close enough.

+++ src/org/broad/igv/ui/panel/ReferenceFrame.java
@@ -544,7 +544,7 @@ public class ReferenceFrame {
      * @param screenPosition
      * @return
      */
-    public double getChromosomePosition(int screenPosition) {
+    public double getChromosomePosition(double screenPosition) {
 
         InsertionMarker i = InsertionManager.getInstance().getSelectedInsertion(getChrName());
